# hdkeychain: Use specialized secp256k1 types.