**This requires #2156**.

This optimizes the various internal implementations that make use of `secp2561` to use the specialized `ModNScalar` and `JacobianPoint` types instead of `big.Int`s.

The following benchmarks show a before and after comparison:

```
benchmark                 old ns/op     new ns/op   delta
-----------------------------------------------------------
BenchmarkDeriveHardened   3702          3547        -4.19%
BenchmarkDeriveNormal     4053          3587        -11.50%
BenchmarkPrivToPub        94.7          94.0        -0.74%
BenchmarkDeserialize      5825          5797        -0.48%
BenchmarkSerialize        9509          9402        -1.13%

benchmark                 old allocs   new allocs   delta
-----------------------------------------------------------
BenchmarkDeriveHardened   14           12           -14.29%
BenchmarkDeriveNormal     15           12           -20.00%
BenchmarkPrivToPub        1            1            +0.00%
BenchmarkDeserialize      4            3            -25.00%
BenchmarkSerialize        3            3            +0.00%

benchmark                 old bytes    new bytes    delta
-----------------------------------------------------------
BenchmarkDeriveHardened   1360         1232         -9.41%
BenchmarkDeriveNormal     1409         1232         -12.56%
BenchmarkPrivToPub        112          112          +0.00%
BenchmarkDeserialize      320          256          -20.00%
BenchmarkSerialize        272          272          +0.00%
```